Created function called histogram
I created a function called histogram in the statistics file. This function accepts the different buckets produced by the buckets function as parameter and calls the sense function for each bucket and adds the returned value to a new named tuple called databuckets which is identical to the buckets except it contains an additional field for the value returned by the sense function for each bucket.