Move python files into new 'sensordesign' folder, change file paths for database, add new file called settings, README in data folder
In order to comply with formatting I created a new folders called 'sensordesign' and 'docs'. The docs folder contains a new file called CONTRIBUTORS which lists contributors. All the python files were moved into the new sensordesign folder but the database kept being created in then top level folder, so the databse path had to use the BASE_DIR base in order to stay in the subfolder. There were problems with circular imports when importing the BASE_DIR from sensordesign.py so a new python file called settings.py was created and the BASE_DIR was defined there. The python files imported BASE_DIR from settings for use in database paths. The data folder README was accidentally deleted so a new one was made with the text from an old commit.